Moving generation code into qirlib
