+++ qirlib/src/generation/mod.rs
@@ -1,0 +1,3 @@
+pub mod emit;
+pub mod interop;
+pub mod qir;
